move unit tests to tests/ integration directory (closes #5)
Splits the 70-test src/tests.rs into 6 concern-focused integration test files
under tests/. Removes #[cfg(test)] mod tests from lib.rs. Public API surface
unchanged (no pub(crate) needed).
